Make tests use SMS RAM and blank rom file
The SMS memory handler was affecting the tests since it was assuming
we were writing to ROM and thus throwing away all writes. Using a blank
rom file at the start and then patching the RAM with the commands and
variables we would like to use makes it so that the tests are ran
correctly.